Did unit tests for the OneTimeCodeService and Integration tests for the methods added to UserService and the RegistrationController

+++ SENG302.Api/Models/Entities/User.cs
@@ -25,7 +25,7 @@ public class User
     [StringLength(6)]
     public string OneTimeCode { get; set; } = String.Empty;
     
-    public int CodeGenerationTime { get; set; }
+    public long CodeGenerationTime { get; set; }
 
     public DateTimeOffset TimeCreated { get; set; }
